Extract optimization opportunity analysis
Move per-method opportunity discovery behind a shared factual resolver while preserving canonical context and allocation interpretation.
